Revert "LPS-81018 Upgrade to the latest Equinox Http Servlet"
This reverts commit 9846d7dc6eb9fc40334664bbeecbcd107dc2b7d2.
