+++ modules/apps/static/portal-osgi-web/portal-osgi-web-wab-extender/build.gradle
@@ -1,6 +1,7 @@
 dependencies {
 	compileInclude group: "com.liferay", name: "javax.servlet.jsp", version: "2.3.3-b02.LIFERAY-PATCHED-7"
 
+	compileOnly fileTree(dir: "lib-patch", exclude: "*-sources.jar", include: "*.jar")
 	compileOnly group: "biz.aQute.bnd", name: "biz.aQute.bndlib", version: "3.1.0"
 	compileOnly group: "com.liferay", name: "org.apache.axis", version: "1.4.LIFERAY-PATCHED-2"
 	compileOnly group: "com.liferay", name: "org.eclipse.osgi", version: "3.10.200-20150904.172142-1.LIFERAY-PATCHED-1"
@@ -13,7 +14,6 @@ dependencies {
 	compileOnly group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
 	compileOnly group: "org.osgi", name: "org.osgi.service.event", version: "1.3.1"
 	compileOnly group: "org.osgi", name: "org.osgi.service.http.whiteboard", version: "1.0.0"
-	compileOnly group: "p2", name: "org.eclipse.equinox.http.servlet", version: "1.5.0.v20180501-1936"
 	compileOnly project(":apps:static:osgi:osgi-util")
 	compileOnly project(":apps:static:portal-configuration:portal-configuration-metatype-api")
 	compileOnly project(":apps:static:portal-osgi-web:portal-osgi-web-servlet-api")
@@ -29,15 +29,10 @@ deployDependencies {
 	from(configurations.compileOnly) {
 		include "commons-fileupload-*.jar"
 		include "commons-io-*.jar"
-		include "org.eclipse.equinox.http.servlet-*.jar"
 	}
-}
 
-repositories {
-	ivy {
-		url "http://download.eclipse.org/eclipse/updates/4.8milestones/S-4.8M7-201805092000/plugins/"
-		layout "pattern", {
-			artifact "[artifact]_[revision](-[classifier])(.[ext])"
-		}
+	from("lib-patch") {
+		exclude "*-sources.jar"
+		include "org.eclipse.equinox.http.servlet-*.jar"
 	}
 }
