+++ modules/apps/static/portal-osgi-web/portal-osgi-web-wab-extender/src/main/java/com/liferay/portal/osgi/web/wab/extender/internal/adapter/HttpAdapter.java
@@ -14,6 +14,7 @@
 
 package com.liferay.portal.osgi.web.wab.extender.internal.adapter;
 
+import com.liferay.portal.kernel.servlet.PortletSessionListenerManager;
 import com.liferay.portal.kernel.util.HashMapDictionary;
 
 import java.lang.reflect.InvocationHandler;
@@ -30,8 +31,13 @@ import javax.servlet.ServletContext;
 import javax.servlet.ServletException;
 import javax.servlet.descriptor.JspConfigDescriptor;
 import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpSession;
+import javax.servlet.http.HttpSessionEvent;
+import javax.servlet.http.HttpSessionListener;
 
 import org.eclipse.equinox.http.servlet.HttpServiceServlet;
+import org.eclipse.equinox.http.servlet.HttpSessionTrackerUtil;
+
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.ServiceRegistration;
 import org.osgi.service.component.ComponentContext;
@@ -117,10 +123,16 @@ public class HttpAdapter {
 				HttpServiceServlet.class.getName(), HttpServlet.class.getName()
 			},
 			_httpServiceServlet, properties);
+
+		PortletSessionListenerManager.addHttpSessionListener(
+			_INVALIDATEHTTPSESSION_LISTENER);
 	}
 
 	@Deactivate
 	protected void deactivate() {
+		PortletSessionListenerManager.removeHttpSessionListener(
+			_INVALIDATEHTTPSESSION_LISTENER);
+
 		_serviceRegistration.unregister();
 
 		_serviceRegistration = null;
@@ -142,6 +154,22 @@ public class HttpAdapter {
 	private static final Class<?>[] _INTERFACES = new Class<?>[] {
 		ServletContext.class
 	};
+
+	private static final HttpSessionListener _INVALIDATEHTTPSESSION_LISTENER =
+		new HttpSessionListener() {
+
+			@Override
+			public void sessionCreated(HttpSessionEvent httpSessionEvent) {
+			}
+
+			@Override
+			public void sessionDestroyed(HttpSessionEvent httpSessionEvent) {
+				HttpSession httpSession = httpSessionEvent.getSession();
+
+				HttpSessionTrackerUtil.invalidate(httpSession.getId());
+			}
+
+		};
 
 	private HttpServiceServlet _httpServiceServlet;
 	private ServiceRegistration<?> _serviceRegistration;
